Moved PreprocessedClusteringAlgorithm out of ClusteringPreprocessor
